requirements.txt
alpaca-py==0.13.1
certifi==2021.10.8
dnspython==1.16.0
pandas==2.1.2
pymongo==3.12.3
python-dotenv==0.19.2
requests==2.31.0
xgboost==2.0.1
lxml==4.9.3
